==> 1 - useful_functions_and_rda.ipynb <==
[Predictive Analytics for CRM Sales Performance](https://www.kaggle.com/datasets/agungpambudi/crm-sales-predictive-analytics)

You will find the dataset here (with other tables). I only used a small part of it.


# This is changed to Raw as there is really no need to run this, but you can change it to Code and run it

import create_pickles
create_pickles.run('files/')

import typing as tp

import IPython.display as ip
import pandas as pd

import pandas_tools as pt_lib


accounts = pd.read_pickle('pickles/accounts.pkl')
sales = pd.read_pickle('pickles/sales.pkl')



accounts_metadata = {
    'name': 'accounts',
    'source_file': 'accounts.csv',
    'source_path': 'Downloads/',
    'source_type': 'CSV'}

accounts.attrs = accounts_metadata

sales_metadata = {
    'name': 'sales',
    'source_file': 'sales_pipeline.csv',
    'source_path': 'Downloads/',
    'source_type': 'CSV'}

sales.attrs = sales_metadata

sales.attrs

pt_lib.tidy_sales(sales)

pt_lib.tidy_accounts(accounts)


# This is an example of merge creating additional rows, columns, and converting employee table to float64
# which is all undesired behaviour
pd.merge(sales, accounts, on='account', how='left')


# This code will generate custom error MergeErrorDuplicateColumns to show
# how merge_safe() is preventing duplication of columns

pandas_tools = pt_lib.PandasTools()

pandas_tools.merge_safe(sales, accounts, on='account')


# After fixing the columns issue by removing 'revenue' column from sales...
# This code will generate custom error MergeErrorDuplicateRows to show
# how merge_safe() is preventing duplication of rows as well

sales_clean = sales.copy()
sales_clean = pandas_tools.exclude_columns(sales_clean, 'revenue')
pandas_tools.merge_safe(sales_clean, accounts, on='account')

# This is good now
sales_clean = sales.copy()
sales_clean = pandas_tools.exclude_columns(sales_clean, 'revenue')
accounts_clean = accounts.copy()
accounts_clean.drop_duplicates(subset='account', inplace=True, ignore_index=True)
pandas_tools.merge_safe(sales_clean, accounts_clean, on='account')

# This code will generate an error to shouw that there is no tools attribute
# of a dataframe
sales_clean.tools

import pandas_tools_with_rda as pd_rda

# Sales with accounts - now we have tools
swat = sales_clean.tools.merge_safe(accounts_clean, on='account')
swat

swat.dtype

swat.dtype.view_floats()

swat.cols.get('close')

swat_metadata = {
    'name': 'sales with accounts',
    'sources' : ['sales', 'accounts'],
    'source_type': 'DataFrame'}

swat.attrs = swat_metadata
swat.to_pickle('pickles/sales_with_accounts.pkl')
