==> 3 - advanced_filtering_and_lineage.ipynb <==
import pandas as pd
import filtering_with_lineage as fl_lib


swat = pd.read_pickle('pickles/swat_clean.pkl')
swat

strategic_review = {
    'series_name': 'strategic_review',
    'statements': (
        (('greater than', 'close_value', 1000), 'big deal focus'),
        (('or', (('lesser than', 'close_value', 100), ('is null', 'close_value'))), 'small deal no focus'),
        (('or', (('greater than', 'employees', 3000), ('equal', 'sector', 'technology'))), 'big_tech')),
    'else': 'further review',
    'dtype': 'string'}

swat.case.else_if_all(strategic_review)


# swat.case.lineage is a list of dictionaries, if I put it in the pd.DataFrame I get a table

pd.DataFrame(swat.case.lineage).astype({'logical_id': 'Int64'})


# This is a second example of IF...ELSEIF of
# quarter. This is just to demonstrate how filters work
# (there are much easier ways to get a quarter)

month_data = swat['close_date'].dt.month
month_data.name = 'month_of_close_date'

to_quarter = {
    'series_name': 'close_quarter',
    'statements': (
        (('lesser than', month_data, 4), 'Q1'),
        (('lesser than', month_data, 7), 'Q2'),
        (('lesser than', month_data, 10), 'Q3')),
    'else': 'Q4',
    'dtype': 'string'}

swat.case.else_if_all(to_quarter)


# This is how it looks like in lineage attribute (list of dictionaries)
swat.case.lineage[:2]



pd.DataFrame(swat.case.lineage).astype({'logical_id': 'Int64'})

```mermaid
---
config:
    theme: forest
---
flowchart LR
    Start["strategic_review"] --> C0
    C0{"close_value > 1000"} -- Yes --> SR0["big deal focus"]
    C0 -- No --> C1
    C1{"close_value < 100<br/>OR<br/>close_value is null ?"}
    C1 -- Yes --> SR1["small deal no focus"]
    C1 -- No --> C2{"employees > 3000 OR sector = technology?"}
    C2 -- Yes --> SR2["big_tech"]
    C2 -- No --> C3{"strategic_review is null ?"}
    C3 -- Yes --> SR3["further review"]
```

```mermaid
---
config:
    theme: forest
---
flowchart LR

    N1 --> D1
    N1 --> D2
    N1 --> D3
    N2 --> D4
    N3 --> D5
    N4 --> D6
    D1 --> O1
    D2 --> O2
    D3 --> O2
    D4 --> O3
    D5 --> O3
    D6 --> O4

    %% outcomes
    subgraph OT1 [strategic_review]
        O1(big deal focus)
        O2(small deal no focus)
        O3(big tech)
        O4(further review)
    end

    %% decision trees
    D1{> 1000}
    subgraph DT1 [OR]
        D2{<100}
        D3{is Null}
    end

    subgraph DT2 [OR]
        D4{> 3000}
        D5{='big tech'}
    end

    D6{is Null}
    
    %% datasources
    subgraph DB1 [sales with accounts]
        N1(close_value)
        N2(employees)
        N3(sector)
        N4(strategic_review)
    end

```

### This is an example of basic mermaid generic flowchart

```mermaid
---
config:
    theme: forest
---
flowchart LR

    N1 --> D1
    D1 -- True --> O1
    D1 -- False --> O2
    N2 --> D2
    D2 -- True --> O3
    D2 -- False --> O4

    %% datasource
    subgraph MT [My Table]
        N1(Column1)
        N2(Column2)
    end
    %% decision
    D1{Rule1}
    D2{Rule2}

    %% outcomes
    O1[(Outcome1)]
    O2[[Outcome2]]
    O3>Outcome3]
    O4{{Outcome4}}
```
